Fix: Storage Refactoring completed
