+++ homepage/package.json
@@ -14,7 +14,8 @@
     "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.12.1",
-    "react-scripts": "^5.0.1"
+    "react-scripts": "^5.0.1",
+    "socket.io-client": "^4.6.2"
   },
   "scripts": {
     "start": "set NODE_ENV = dev&&react-scripts start",
